requirements.txt
fastapi
uvicorn
jinja2
python-multipart
websockets
google-genai
